Style: PersonnelContext의 hook을 파일로 분리
